Move model mixins to separate submodule
